refactor: move encoding guards to schema
